requirements.txt
pycryptodome==3.19.0
bcrypt==4.1.2 